+++ rclpy_message_converter/CHANGELOG.rst
@@ -9,8 +9,8 @@ Change Log
 * Add NestedUint8ArrayTestService tests
 * propagate check_types in _convert_to_ros_type (`#51 <https://github.com/uos/rospy_message_converter/issues/51>`_)
   Co-authored-by: Martin Günther <martin.guenther@dfki.de>
-* Fix binary_array_as_bytes=False with nested msgs
-* Add param binary_array_as_bytes
+* Fix base64_encoding=False with nested msgs
+* Add param base64_encoding
   Closes `#45 <https://github.com/uos/rospy_message_converter/issues/45>`_.
 * Contributors: Marc Bosch-Jorge, Martin Günther, Otacon5555